test(content-mapper): cover declaration map type queries
